starlinks added to db, juniper setup started

+++ migrations/2021-07-09-202759_rockets/up.sql
@@ -6,7 +6,7 @@ CREATE TABLE rockets
   active Boolean,
   stages Int,
   boosters Int,
-  cost_per_launch bigint,
+  cost_per_launch Int,
   success_rate_pct Int,
   first_flight character varying,
   country character varying,
